♻️ 일정 관리 페이지 리팩토링
